requirements.txt
streamlit>=1.31.0
requests
speechrecognition>=3.10.0
streamlit-mic-recorder
google-generativeai
